requirements.txt
alembic==0.9.3
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-MySQLdb==0.2.0
flask-restaction==0.25.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.10
PyJWT==1.5.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.1
simple-yaml==0.1.0
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.21.1
validr==0.14.1
Werkzeug==0.12.2
